Member microservice updated

+++ src/service/java/nz/paymark/member/dao/impl/MemberDaoImpl.java
@@ -1,12 +1,13 @@
 package nz.paymark.member.dao.impl;
 
-import static nz.paymark.shared.validation.util.ExceptionThrower.throwNotFoundIf;
-import static nz.paymark.shared.validation.util.ExceptionThrower.throwNullArgumentFor;
+import static nz.paymark.shared.models.validation.util.GeneralExceptionThrower.throwNullArgumentFor;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 
+import javax.annotation.Resource;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
@@ -15,13 +16,14 @@ import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
 
+import nz.paymark.client.shared.web.exception.RecordNotFoundException;
 import nz.paymark.logging.LoggerFactory;
 import nz.paymark.member.dao.MemberDao;
 import nz.paymark.member.model.Member;
 import nz.paymark.member.model.MemberSearchCriteria;
-//import nz.paymark.member.model.Member_;
 import nz.paymark.member.model.Member_;
-import nz.paymark.member.model.enumerator.MemberStatus;
+import nz.paymark.member.model.enumerator.MemberRoles;
+import nz.paymark.shared.models.validation.SimpleValidator;
 
 import org.slf4j.Logger;
 import org.springframework.stereotype.Repository;
@@ -29,35 +31,52 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Repository
 public class MemberDaoImpl implements MemberDao {
+	
+	@Resource
+    private SimpleValidator validator;
 
 	public static Logger logger = LoggerFactory.getLogger();
 	
-	@PersistenceContext private EntityManager em;
+	@PersistenceContext
+	private EntityManager em;
+	
+	@Override
+    public boolean touch() {
+        CriteriaBuilder cb = em.getCriteriaBuilder();
+        CriteriaQuery<Long> cq = cb.createQuery(Long.class);
+        cq.select(cb.count(cq.from(Member.class)));
+        return em.createQuery(cq).getSingleResult() > -1;
+    }
+	
+	@Override
+    public void validate(Member member) {
+        validator.validate(member);
+    }
 
 	@Override
 	@Transactional(readOnly=false)
 	public Member createMember(Member member) {
 		throwNullArgumentFor(member);		
-		
-		return em.merge(member);
+		LocalDateTime now = LocalDateTime.now();
+        member.setCreationTime(now);
+        member.setModifiedTime(now);
+        validate(member);
+        return em.merge(member);
 	}
 	
 	@Override
 	@Transactional(readOnly=false)
 	public Member updateMember(Member member) {
 		throwNullArgumentFor(member);
+		validate(member);
 		return em.merge(member);
 	}
 
 	@Override
 	@Transactional(readOnly=true)
-	public Optional<Member> findMemberById(String id) {
+	public Member getMember(String id) {
 		throwNullArgumentFor(id);
-		Member member = em.find(Member.class, id);
-		if(null == member){
-			return Optional.empty();
-		}
-		return Optional.of(member);
+		return em.find(Member.class, id);
 	}
 
 
@@ -66,13 +85,14 @@ public class MemberDaoImpl implements MemberDao {
 	public void deleteMember(String id) {
 		throwNullArgumentFor(id);
         Member member = em.find(Member.class, id);
-        throwNotFoundIf(member == null);
+        if (member == null) {
+            throw new RecordNotFoundException();
+        }
         em.remove(member);  		
 	}
 
 	@Override
-	public List<Member> searchMembers(MemberSearchCriteria searchCriteria) {
-
+	public List<Member> searchMembersByCriteria(MemberSearchCriteria searchCriteria) {
 		CriteriaBuilder criteriaBuilder = em.getCriteriaBuilder();
 		CriteriaQuery<Member> criteriaQuery = criteriaBuilder
 				.createQuery(Member.class);
@@ -80,32 +100,86 @@ public class MemberDaoImpl implements MemberDao {
 		Root<Member> memberRoot = criteriaQuery.from(Member.class);
 
 		List<Predicate> conditions = new ArrayList<>();
-		String role  = searchCriteria.getRole();
-		if (role != null && role.length() != 0) {
-			conditions.add(criteriaBuilder.equal(memberRoot.get(Member_.role),
-					role));
-		}
 		
-		MemberStatus status = searchCriteria.getStatus();
-		if (status != null) {
-			conditions.add(criteriaBuilder.equal(memberRoot.get(Member_.status),
-					status));
-		}
-		
-		String organisationId = searchCriteria.getOrganisationId();
-		if (organisationId != null && organisationId.length() != 0) {
-			conditions.add(criteriaBuilder.equal(memberRoot.get(Member_.organisationId),
-					organisationId));
-		}
-		
-		String userId = searchCriteria.getUserId();
-		if (userId != null && userId.length() != 0) {
-			conditions.add(criteriaBuilder.equal(memberRoot.get(Member_.userId),
-					userId));
+		if (null != searchCriteria) {
+			String id = searchCriteria.getId();
+            if (id != null && id.length() != 0) {
+                conditions.add(criteriaBuilder.equal(
+                        memberRoot.get(Member_.id), id));
+            }
+            
+            String userId = searchCriteria.getUserId();
+            if(userId != null && userId.length() != 0){
+            	conditions.add(criteriaBuilder.equal(
+                        memberRoot.get(Member_.userId), userId));
+            }
+            
+            String organisationId = searchCriteria.getOrganisationId();
+            if(organisationId != null && organisationId.length() != 0){
+            	conditions.add(criteriaBuilder.equal(
+                        memberRoot.get(Member_.organisationId), organisationId));
+            }
+            
+            MemberRoles role  = searchCriteria.getRole();
+			if (role != null) {
+				conditions.add(criteriaBuilder.equal(memberRoot.get(Member_.role),
+						role));
+			}
+			
+			LocalDateTime cTimeBegin = searchCriteria.getCreationTimeBegin();
+	        LocalDateTime cTimeEnd = searchCriteria.getCreationTimeEnd();
+	
+	        if (cTimeBegin != null && cTimeEnd == null) {
+	            conditions.add(criteriaBuilder.greaterThanOrEqualTo(
+	            		memberRoot.get(Member_.creationTime), cTimeBegin));
+	        } else if (cTimeBegin == null && cTimeEnd != null) {
+	            conditions.add(criteriaBuilder.lessThanOrEqualTo(
+	            		memberRoot.get(Member_.creationTime), cTimeEnd));
+	        } else if (cTimeBegin != null && cTimeEnd != null) {
+	            conditions.add(criteriaBuilder.between(
+	            		memberRoot.get(Member_.creationTime), cTimeBegin,
+	                    cTimeEnd));
+	        }
+	
+	        LocalDateTime mTimeBegin = searchCriteria
+	                .getModificationTimeBegin();
+	        LocalDateTime mTimeEnd = searchCriteria.getModificationTimeEnd();
+	
+	        if (mTimeBegin != null && mTimeEnd == null) {
+	            conditions.add(criteriaBuilder.greaterThanOrEqualTo(
+	            		memberRoot.get(Member_.modifiedTime), mTimeBegin));
+	        } else if (mTimeBegin == null && mTimeEnd != null) {
+	            conditions.add(criteriaBuilder.lessThanOrEqualTo(
+	            		memberRoot.get(Member_.modifiedTime), mTimeEnd));
+	        } else if (mTimeBegin != null && mTimeEnd != null) {
+	            conditions.add(criteriaBuilder.between(
+	            		memberRoot.get(Member_.modifiedTime), mTimeBegin,
+	                    mTimeEnd));
+	        }
 		}
 		criteriaQuery.where(conditions.toArray(new Predicate[] {}));
 
 		TypedQuery<Member> query = em.createQuery(criteriaQuery);
 		return query.getResultList();
 	}
+	
+	@Override
+    public List<Member> searchMembers(List<String> ids) {
+
+        if (ids == null || ids.size() == 0) {
+            return Collections.emptyList();
+        }
+
+        CriteriaBuilder criteriaBuilder = em.getCriteriaBuilder();
+        CriteriaQuery<Member> criteriaQuery = criteriaBuilder
+                .createQuery(Member.class);
+
+        Root<Member> member = criteriaQuery
+                .from(Member.class);
+        criteriaQuery.select(member).where(
+                member.get(Member_.id).in(ids));
+
+        TypedQuery<Member> query = em.createQuery(criteriaQuery);
+        return query.getResultList();
+    }
 }
